+++ src/sbbs3/cmdshell.h
@@ -1,9 +1,4 @@
-/* cmdshell.h */
-// vi: tabstop=4
-
 /* Synchronet command shell/module constants and structure definitions */
-
-/* $Id: cmdshell.h,v 1.16 2020/04/16 07:39:46 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -18,20 +13,8 @@
  * See the GNU General Public License for more details: gpl.txt or			*
  * http://www.fsf.org/copyleft/gpl.html										*
  *																			*
- * Anonymous FTP access to the most recent released source is available at	*
- * ftp://vert.synchro.net, ftp://cvs.synchro.net and ftp://ftp.synchro.net	*
- *																			*
- * Anonymous CVS access to the development source and modification history	*
- * is available at cvs.synchro.net:/cvsroot/sbbs, example:					*
- * cvs -d :pserver:anonymous@cvs.synchro.net:/cvsroot/sbbs login			*
- *     (just hit return, no password is necessary)							*
- * cvs -d :pserver:anonymous@cvs.synchro.net:/cvsroot/sbbs checkout src		*
- *																			*
  * For Synchronet coding style and modification guidelines, see				*
  * http://www.synchro.net/source.html										*
- *																			*
- * You are encouraged to submit any modifications (preferably in Unix diff	*
- * format) via e-mail to mods@synchro.net									*
  *																			*
  * Note: If this box doesn't appear square, then you need to fix your tabs.	*
  ****************************************************************************/
@@ -675,11 +658,11 @@ typedef struct {					/* Command shell image */
 	uint32_t
 			*str_var_name,			/* String variable names (CRC-32) */
 			*int_var_name;			/* Integer variable names (CRC-32) */
-	long	retval, 				/* Return value */
+	int		retval, 				/* Return value */
 			misc,					/* Misc bits */
 			switch_val; 			/* Current switch value */
 
-	long	length; 				/* Length of image */
+	size_t	length; 				/* Length of image */
 
 	} csi_t;
